Fix missing type casting

+++ core/src/com/shootan/stellarpatrol/gameobjects/GameObject.java
@@ -79,7 +79,7 @@ public abstract class GameObject {
     }
 
     private float getAspectRatio() {
-        return textureRegion.getRegionWidth() / textureRegion.getRegionHeight();
+        return (float) textureRegion.getRegionWidth() / textureRegion.getRegionHeight();
     }
 
     private float getTextureRotation() {
